analysis/backtests/requirements.txt
# CHAOS Token Backtesting Dependencies

# Data fetching
requests>=2.31.0
pandas>=2.1.0
numpy>=1.24.0

# APIs
blockfrost-python>=0.5.3
pycoingecko>=3.1.0

# Analysis
matplotlib>=3.7.0
seaborn>=0.12.0
scipy>=1.11.0

# Backtesting
backtrader>=1.9.78.123

# Performance metrics
empyrical>=0.5.5

# Utilities
python-dotenv>=1.0.0
tqdm>=4.66.0
